Fixed warnings reported by the IntelliJ IDEA Code Inspector in module st-exerciseviewer.

+++ st-exerciseviewer/src/main/java/de/saring/exerciseviewer/parser/AbstractExerciseParser.java
@@ -118,7 +118,7 @@ public abstract class AbstractExerciseParser implements ExerciseParser {
             timeSplitBefore = lap.getTimeSplit();
 
             lap.getSpeed().setSpeedAVG(CalculationUtils.calculateAvgSpeed(
-                    lapDistance / 1000f, (int) Math.round(lapDuration / 10f)));
+                    lapDistance / 1000f, Math.round(lapDuration / 10f)));
         }
     }
 }
